Move withDirLock into new other-module
Then we have smaller non-domain specific public API surface: that's good.
